Refactor common methods into a Content\TestCase class
